Move loader_directory from tools to tool_util [WIP]
This is WIP: this move (obviously) breaks many things, whick are
currently being fixed.
