+++ lib/galaxy/tool_util/loader_directory.py
