+++ lib/galaxy/tools/toolbox/base.py
@@ -12,7 +12,7 @@ from six.moves.urllib.parse import urlparse
 
 from galaxy.exceptions import MessageException, ObjectNotFound
 from galaxy.tool_util.deps import build_dependency_manager
-from galaxy.tools.loader_directory import looks_like_a_tool
+from galaxy.tool_util.loader_directory import looks_like_a_tool
 from galaxy.util import (
     ExecutionTimer,
     listify,
